split design files to match existing controllers + required 'type' param
